+++ src/core/index.ts
@@ -10,3 +10,4 @@ export * from "./builders/medication-statement.js";
 export * from "./builders/encounter.js";
 export * from "./builders/diagnostic-report.js";
 export * from "./builders/bundle.js";
+export * from "./recipes.js";